pyproject.toml
[tool.poetry]
name = "homework33"
version = "0.1.0"
description = ""
authors = ["Nikalh <nikalhov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.0.1"

docker-compose = "^1.29.2"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
django-restframework = "^0.0.1"
djangorestframework = "^3.14.0"
social-auth-app-django = "^5.0.0"
django-extensions = "^3.2.1"
envparse = "^0.2.0"
python-dotenv = "0.13.0"
django-filter = "^22.1"
pydantic = "^1.10.6"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"
djangorestframework-simplejwt = "^5.2.2"
faker = "^18.3.1"



[tool.poetry.group.dev.dependencies]
pre-commit = "^3.0.4"
ansible-vault-win = "^1.0.0"
django-extensions = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
